+++ cardano-node/src/Cardano/Node/Configuration/POM.hs
@@ -21,9 +21,11 @@ module Cardano.Node.Configuration.POM
   )
 where
 
+import           Control.Monad (when)
 import           Data.Aeson
 import qualified Data.Aeson.Types as Aeson
 import           Data.Bifunctor (Bifunctor (..))
+import           Data.Maybe (isJust)
 import           Data.Monoid (Last (..))
 import           Data.Text (Text)
 import qualified Data.Text as Text
@@ -49,6 +51,16 @@ import           Ouroboros.Consensus.Mempool (MempoolCapacityBytes (..),
 import qualified Ouroboros.Consensus.Node as Consensus (NetworkP2PMode (..))
 import           Ouroboros.Consensus.Storage.LedgerDB.DiskPolicy (SnapshotInterval (..))
 import           Ouroboros.Network.NodeToNode (AcceptedConnectionsLimit (..), DiffusionMode (..))
+
+-- | Parse field that have been removed from the configuration file and
+-- fail if they are present.
+--
+-- This is used to notify users that a field has been removed from the
+-- configuration file.
+failOnRemovedField :: Aeson.Object -> Key -> String -> Aeson.Parser ()
+failOnRemovedField obj removedField errorMessage = do
+  mVal :: Maybe Aeson.Value <- obj .:? removedField
+  when (isJust mVal) $ fail errorMessage
 
 data NetworkP2PMode = EnabledP2PMode | DisabledP2PMode
   deriving (Eq, Show, Generic)
@@ -383,9 +395,12 @@ instance FromJSON PartialNodeConfiguration where
              }
 
       parseHardForkProtocol v = do
-        npcTestEnableDevelopmentHardForkEras
-          <- v .:? "TestEnableDevelopmentHardForkEras"
-               .!= False
+
+        npcExperimentalHardForksEnabled <- do
+          failOnRemovedField v "TestEnableDevelopmentHardForkEras"
+            "TestEnableDevelopmentHardForkEras has been renamed to ExperimentalHardForksEnabled"
+
+          v .:? "ExperimentalHardForksEnabled" .!= False
 
         npcTestShelleyHardForkAtEpoch   <- v .:? "TestShelleyHardForkAtEpoch"
         npcTestShelleyHardForkAtVersion <- v .:? "TestShelleyHardForkAtVersion"
@@ -405,27 +420,27 @@ instance FromJSON PartialNodeConfiguration where
         npcTestConwayHardForkAtEpoch   <- v .:? "TestConwayHardForkAtEpoch"
         npcTestConwayHardForkAtVersion <- v .:? "TestConwayHardForkAtVersion"
 
-        pure NodeHardForkProtocolConfiguration {
-               npcTestEnableDevelopmentHardForkEras,
+        pure NodeHardForkProtocolConfiguration
+          { npcExperimentalHardForksEnabled
 
-               npcTestShelleyHardForkAtEpoch,
-               npcTestShelleyHardForkAtVersion,
+          , npcTestShelleyHardForkAtEpoch
+          , npcTestShelleyHardForkAtVersion
 
-               npcTestAllegraHardForkAtEpoch,
-               npcTestAllegraHardForkAtVersion,
+          , npcTestAllegraHardForkAtEpoch
+          , npcTestAllegraHardForkAtVersion
 
-               npcTestMaryHardForkAtEpoch,
-               npcTestMaryHardForkAtVersion,
+          , npcTestMaryHardForkAtEpoch
+          , npcTestMaryHardForkAtVersion
 
-               npcTestAlonzoHardForkAtEpoch,
-               npcTestAlonzoHardForkAtVersion,
+          , npcTestAlonzoHardForkAtEpoch
+          , npcTestAlonzoHardForkAtVersion
 
-               npcTestBabbageHardForkAtEpoch,
-               npcTestBabbageHardForkAtVersion,
+          , npcTestBabbageHardForkAtEpoch
+          , npcTestBabbageHardForkAtVersion
 
-               npcTestConwayHardForkAtEpoch,
-               npcTestConwayHardForkAtVersion
-             }
+          , npcTestConwayHardForkAtEpoch
+          , npcTestConwayHardForkAtVersion
+          }
 
 -- | Default configuration is mainnet
 defaultPartialNodeConfiguration :: PartialNodeConfiguration