+++ cardano-testnet/src/Testnet/Babbage.hs
@@ -97,7 +97,7 @@ babbageTestnet testnetOptions H.Conf {..} = do
     . HM.insert "TestMaryHardForkAtEpoch" (toJSON @Int 0)
     . HM.insert "TestAlonzoHardForkAtEpoch" (toJSON @Int 0)
     . HM.insert "TestBabbageHardForkAtEpoch" (toJSON @Int 0)
-    . HM.insert "TestEnableDevelopmentHardForkEras" (toJSON True)
+    . HM.insert "ExperimentalHardForksEnabled" (toJSON True)
     . flip HM.alter "setupScribes"
         ( fmap
           . J.rewriteArrayElements