+++ cardano-testnet/src/Testnet/Cardano.hs
@@ -256,7 +256,7 @@ cardanoTestnet testnetOptions H.Conf {..} = do
     . HM.insert "LastKnownBlockVersion-Minor" (J.toJSON @Int 0)
     . HM.insert "TraceBlockchainTime" (J.toJSON True)
     . HM.delete "GenesisFile"
-    . HM.insert "TestEnableDevelopmentHardForkEras" (J.toJSON @Bool True)
+    . HM.insert "ExperimentalHardForksEnabled" (J.toJSON @Bool True)
     . HM.insert "EnableP2P" (J.toJSON @Bool (cardanoEnableP2P testnetOptions))
     . flip HM.alter "setupScribes"
         ( fmap